+++ sneakctl_server/flaskr/routes/db.py
@@ -2,7 +2,7 @@ from flask import Blueprint
 
 from sneakctl_server.flaskr.db import get_db
 
-db_blueprint = Blueprint('blueprint', __name__, url_prefix='/db')
+db_blueprint = Blueprint('db_blueprint', __name__, url_prefix='/db')
 
 
 @db_blueprint.route('/test', methods=['GET'])